Greatly simplified graph serialization
I realized I don't have to serialize the nodes in any particular order
since the child lists are not finalized until they're ALL deserialized.

As part of this commit, the Graph class in RTGUI changed quite a lot.
The "RootValues" and "ExtraNodes" collections are now private to reduce
the chance of using the class poorly.